+++ test/timepicker_test.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import DatePicker from "../src/index.jsx";
-import { newDate, formatDate } from "../src/date_utils.js";
+import { newDate, formatDate } from "../src/date_utils.ts";
 import { getKey } from "./test_utils.js";
 import { render, fireEvent, waitFor } from "@testing-library/react";
