+++ test/year_dropdown_test.test.js
@@ -2,7 +2,7 @@ import React from "react";
 import range from "lodash/range";
 import YearDropdown from "../src/year_dropdown.jsx";
 import { render, fireEvent } from "@testing-library/react";
-import { newDate } from "../src/date_utils.js";
+import { newDate } from "../src/date_utils.ts";
 
 describe("YearDropdown", () => {
   let yearDropdown;
