+++ test/year_picker_test.test.js
@@ -2,7 +2,7 @@ import React from "react";
 import DatePicker from "../src/index.jsx";
 import Year from "../src/year.jsx";
 import { render, fireEvent } from "@testing-library/react";
-import * as utils from "../src/date_utils.js";
+import * as utils from "../src/date_utils.ts";
 import { getKey } from "./test_utils.js";
 
 describe("YearPicker", () => {
